client/requirements.txt
websocket-client
pystray
Pillow
requests
